+++ changelog.md
@@ -1,2 +1,5 @@
+# Features
+- 1.21.3
+
 # Fixes
 - item frames missing texture